Restructured and modularized schema types
